Enrich flows with K8S metadata
For now, just in sflow. This is just a hack, to properly write this
we need to create an Enricher interface and add it to StateSFlow.
Then we write a K8S enricher and that way we open the door to other
enrichers.
I also foresee adding a flag to enable the enrichers.
Finally, the K8S fields should be added to a new var, instead of shoving
it onto defaultFields.